# CORE-376 kafka/quota_mgr: fix client quota throttling to be node-wide