… src/v/kafka/server/quota_manager.h
… src/v/kafka/server/snc_quota_manager.h
… src/v/kafka/server/tests/atomic_token_bucket_test.cc
… src/v/kafka/server/tests/quota_manager_bench.cc
… src/v/kafka/server/tests/quota_manager_test.cc
… src/v/kafka/server/tests/snc_quota_manager_test.cc
… src/v/kafka/server/tests/token_bucket_algorithm_test.cc
… src/v/kafka/server/token_bucket_rate_tracker.h
… src/v/redpanda/application.cc
… src/v/redpanda/application.h
… src/v/resource_mgmt/rate.h
… src/v/ssx/include/ssx/aligned.h
… src/v/ssx/include/ssx/sharded_ptr.h
… src/v/ssx/include/ssx/sharded_value.h
… src/v/ssx/tests/sharded_ptr_test.cc
… tests/rptest/tests/cluster_quota_test.py