@pgellert 2024-04-30
> Where are configurations for groups watched and updated? Does that need to be limited to the worker shard? Does it need to copy the map and deal with the additions and removals?

Good question. It looks like we don't watch for config changes but only apply them when we first see a new client id (by always reading the latest config values). That's problematic for long-living client ids. I think we should fix that and actively watch for config changes, but I also think that it should be a separate PR. I'll make a ticket for it if it sounds good to you too.