@pgellert 2024-05-17
And [another small one](https://github.com/redpanda-data/redpanda/compare/4abc93be3baac29af3434dd451e1b3cd74c3df91..e727cf97f57ac2c6278c1680efc4c9095e4f17a2) to fix the linting error.