@pgellert 2024-05-17
[Rebased to dev](https://github.com/redpanda-data/redpanda/compare/e727cf97f57ac2c6278c1680efc4c9095e4f17a2..6769a080a6510448a090be43c646b6a7565eb736). I'll address the code review above next.