@pgellert 2024-05-20
[Force-pushed](https://github.com/redpanda-data/redpanda/compare/6769a080a6510448a090be43c646b6a7565eb736..ad0fc7434411b366ab17552e7d21109befd5af96) with more code review changes