@pgellert 2024-05-21
[Force-pushed](https://github.com/redpanda-data/redpanda/compare/ad0fc7434411b366ab17552e7d21109befd5af96..10d3124d036be41ac25ff7b79b95b8882c508415) with a microbenchmark in the last commit and some bug fixes.

#### Benchmark
I've now added a microbenchmark to the fetch-related methods of quota manager to try to measure the worst-case performance of these methods under high cross-core contention. The worst case seems to be when all the client ids are the same and so all the cores are contending on the atomic accesses to the client quotas. Even in this case we could handle ~10M requests in ~1.427s which is about 136ns per request. This is certainly more than what shard-local access would be, but it is still a relatively low number that should be within our performance budget for broker-wide client quota tracking.

```
single run iterations:    0
single run duration:      1.000s
number of runs:           5
number of cores:          32
random seed:              4113225649

test                                                iterations      median         mad         min         max      allocs       tasks        inst
quota_group.test_quota_manager_on_unlimited_shared           1      1.427s     8.987ms      1.410s      1.436s 1638484.800    2360.200         0.0
quota_group.test_quota_manager_on_unlimited_unique           6   221.734ms     9.034ms   209.158ms   230.804ms 2034272.233    9752.833         0.0
quota_group.test_quota_manager_on_limited_shared             1      1.156s     1.515ms      1.153s      1.158s 1638485.000    1590.600         0.0
quota_group.test_quota_manager_on_limited_unique             6   187.928ms     1.592ms   186.336ms   229.697ms 2033969.700    9680.200         0.0
quota_group.test_quota_manager_off_shared                    1      1.254s     2.329ms      1.252s      1.268s 1638484.200    2773.200         0.0
quota_group.test_quota_manager_off_unique                    6   165.436ms     5.238ms   160.198ms   214.782ms 2034187.600    9702.833         0.0
```

#### Bug fix
Furthermore, in this force push, I also fixed a bug to do with the lifetime of the `client_quota` objects. We need to ensure that the `client_quota` objects returned from the map are safe to hold on to across scheduling points and that even if the `client_quota` is removed from the map during `gc()`, the last desctruction of the `client_quota` object should happen on shard 0. For this, I've introduced a reference wrapped (`value_node`) to ensure exactly this using `std::shared_ptr` and `ss::foreign_ptr`.