@pgellert 2024-05-22
[Force-pushed](https://github.com/redpanda-data/redpanda/compare/10d3124d036be41ac25ff7b79b95b8882c508415..25227e5829985a2e05ad9c16a5ef5642a1a5ffa6) further changes based on code review and local testing:
* The partition mutation quotas now also use the atomic token buckets. This also required some changes to the tests (see the commit message).
* I've changed all uses of the `atomic_token_bucket` to be with `capped_release::no`, because we don't need it. After some testing I realized that we need `capped_release::no` since we don't need to hold onto units from the token bucket and restrict the releasing of tokens back to the token bucket. I was prompted to do this by the partition mutation ducktape tests as I noticed that the token bucket wasn't getting replenished because we never called `release()` on the token capped token bucket. I've now added more tests around both the `atomic_token_bucket` and `quota_manager` to exercise more of the behaviour.
* I've added a check into the throttling methods to only insert into the quotas map if throttling is configured for the client. This helps with the case when quotas are turned off, as shown by the updated benchmark (when quotas are off, quota management has a ~2.5ns overhead.
* Deleted a bunch of unused code 🔥