@pgellert 2024-05-22
[Force-pushed](https://github.com/redpanda-data/redpanda/compare/25227e5829985a2e05ad9c16a5ef5642a1a5ffa6..d217383b5285c8c0c868bd0d2f010d5b6e525192) with changes to avoid the `sharded_ptr` copy on the request path and avoid the `value_node` in the map. Improved benchmark:
```
test                                                iterations      median         mad         min         max      allocs       tasks        inst
quota_group.test_quota_manager_on_unlimited_shared           3   457.729ms   171.802us   457.558ms   458.047ms 2293843.600     685.067         0.0
quota_group.test_quota_manager_on_unlimited_unique           5   192.070ms     3.564ms   186.680ms   198.077ms 2032395.040    9758.880         0.0
quota_group.test_quota_manager_on_limited_shared             3   474.613ms   305.042us   473.773ms   474.918ms 2293843.200     702.533         0.0
quota_group.test_quota_manager_on_limited_unique             6   197.639ms     2.688ms   193.524ms   200.552ms 2032468.133    9730.600         0.0
quota_group.test_quota_manager_off_shared                   41    24.975ms   195.504us    24.780ms    28.671ms 1638437.693      58.068         0.0
quota_group.test_quota_manager_off_unique                   49    20.691ms   254.579us    20.418ms    21.071ms  655397.441      47.833         0.0
```