@pgellert 2024-05-23
[Force-pushed](https://github.com/redpanda-data/redpanda/compare/d217383b5285c8c0c868bd0d2f010d5b6e525192..954c5f981d11b076987a8f7c83f6af8ab668cafb) to adopt the suggestions above.