@pgellert 2024-05-24
[Force-pushed](https://github.com/redpanda-data/redpanda/compare/954c5f981d11b076987a8f7c83f6af8ab668cafb..f7dbb46559f4ff14c732f1bf4773985fe1f92143) with more suggested changes. I have yet to benchmark and address the point about adding padding around the token buckets.