@vbotbuildovich 2024-05-24
<!--NEW FAILURES-->
new failures in https://buildkite.com/redpanda/redpanda/builds/49526#018fab34-2027-4ad1-a6ed-9e84e8cfe928:
```
"rptest.tests.usage_test.UsageTestCloudStorageMetrics.test_usage_manager_cloud_storage"
```
<!--8d5d19a8-ffea-4fba-add2-4569e2642cd6-->