@pgellert 2024-05-28
[Force-pushed](https://github.com/redpanda-data/redpanda/compare/f7dbb46559f4ff14c732f1bf4773985fe1f92143..7371aaf23d0bcf962b8b8f8e1df226f523c9477a) with:
* Addressing the comment about passing the argument by `const&` to the close of the `sharded_ptr::update` function
* Added padding around the token buckets to align them to the cache line as the benchmarks confirmed that this leads to a performance improvement when multiple cores contend on neighbouring token buckets. This leads to ~25% improvement in the pathological case of having 16-16 cores updating the neighbouring fetch and produce token buckets.

```
Without cache alignment
test                                                iterations      median         mad         min         max      allocs       tasks        inst
quota_group.test_quota_manager_on_unlimited_shared           3   402.589ms     4.656ms   361.566ms   407.246ms 2293843.800    1231.467         0.0
quota_group.test_quota_manager_on_unlimited_unique           7   165.404ms     1.589ms   162.938ms   169.582ms 2033037.571    9595.000         0.0
quota_group.test_quota_manager_on_limited_shared             3   343.729ms   805.020us   342.924ms   346.064ms 2293843.467    1065.067         0.0
quota_group.test_quota_manager_on_limited_unique             7   164.961ms   276.672us   164.685ms   166.071ms 2033145.971    9587.514         0.0
quota_group.test_quota_manager_off_shared                    5   223.623ms     1.085ms   222.348ms   225.867ms 1639174.400     208.880         0.0
quota_group.test_quota_manager_off_unique                    9   111.680ms   719.302us   110.740ms   112.762ms 1448054.867    9181.511         0.0

With cache alignment
test                                                iterations      median         mad         min         max      allocs       tasks        inst
quota_group.test_quota_manager_on_unlimited_shared           4   302.860ms   869.473us   301.991ms   306.604ms 2293843.950    1022.550         0.0
quota_group.test_quota_manager_on_unlimited_unique           7   161.132ms   975.600us   160.103ms   164.485ms 2032651.086    9590.971         0.0
quota_group.test_quota_manager_on_limited_shared             4   307.775ms   356.912us   304.962ms   308.553ms 2293843.500    1089.250         0.0
quota_group.test_quota_manager_on_limited_unique             7   160.303ms     1.421ms   155.586ms   169.324ms 2032743.971    9594.314         0.0
quota_group.test_quota_manager_off_shared                    5   248.071ms     3.099ms   231.620ms   256.416ms 1639174.360     207.240         0.0
quota_group.test_quota_manager_off_unique                    6   162.579ms    10.868ms   149.720ms   256.007ms 1448036.300    9227.067         0.0
```